feat(mirror tool sub items): added working sub items for mirror tools - added sub items - adding toggling of mirror tool - hooked up events etc, all works
